requirements.txt
requests
playwright
colorama
tqdm
python-dotenv
apprise
cryptography
